requirements.txt
3==0.0.0
asgiref==3.4.1
boto3==1.20.13
botocore==1.23.13
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.8
django-ckeditor==5.9.0
django-dotenv==1.4.2
django-generate-secret-key==1.0.2
django-js-asset==1.2.2
django-storages==1.12.3
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
Pillow==8.3.2
psycopg2==2.9.2
psycopg2-binary==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.7
version==0.1.0
